requirements.txt
python_telegram_bot==13.7
beautifulsoup4==4.10.0
python-dotenv==0.19.0
telegram==0.0.1
telegram_send==0.25
